refactor: split trade contract, dialogue, diplomacy, and shell dialog parts (#3878)
Further modularize large trade-screen contract keys, call-to-arms row
widgets, diplomacy detail card/section parts, and leader-selection slot
rows to keep files under the repo size target without changing public APIs.

Refs #3878
