+++ app/lib/features/game/screens/diplomacy/diplomacy_detail_screen_widgets_cards.dart
@@ -1,0 +1,69 @@
+part of 'diplomacy_detail_screen.dart';
+
+/// Inner heading inside a `_DetailCard`. Matches the GAME30002 mockup
+/// `.card h3` rule (display font, 13 px, `--muted`, uppercase, letter-spacing
+/// 0.06 em). Separate from `CtSectionLabel` because card titles do not paint
+/// a bottom border on the mockup.
+class _CardTitle extends StatelessWidget {
+  const _CardTitle(this.text);
+
+  final String text;
+
+  @override
+  Widget build(BuildContext context) {
+    final ThemeData theme = Theme.of(context);
+    final TextStyle baseDisplay =
+        theme.textTheme.titleSmall ?? const TextStyle(fontSize: 14);
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 8),
+      child: Text(
+        text.toUpperCase(),
+        style: baseDisplay.copyWith(
+          color: EditorialMonoclePalette.muted,
+          fontSize: 13,
+          fontWeight: FontWeight.w700,
+          letterSpacing: 0.06 * 13,
+        ),
+      ),
+    );
+  }
+}
+
+/// Framed card with gradient background (`--surface-lite → --surface →
+/// --bg-deep`) and a 1 px `--border` outline. Mirrors mockup `.card`.
+class _DetailCard extends StatelessWidget {
+  const _DetailCard({required this.title, required this.child});
+
+  final String title;
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return DecoratedBox(
+      decoration: BoxDecoration(
+        gradient: LinearGradient(
+          begin: Alignment.topCenter,
+          end: Alignment.bottomCenter,
+          stops: const <double>[0.0, 0.3, 1.0],
+          colors: <Color>[
+            EditorialMonoclePalette.surfaceLite,
+            EditorialMonoclePalette.surface,
+            EditorialMonoclePalette.bgDeep,
+          ],
+        ),
+        border: Border.all(color: EditorialMonoclePalette.border, width: 1),
+      ),
+      child: Padding(
+        padding: const EdgeInsets.all(14),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            _CardTitle(title),
+            child,
+          ],
+        ),
+      ),
+    );
+  }
+}
