+++ app/lib/features/shell/new_game_leader_selection_dialog.dart
@@ -20,6 +20,7 @@ import 'package:colonizethis_app/widgets/gp_default_map_color_swatch.dart';
 
 part 'new_game_leader_selection_dialog_state_base.dart';
 part 'new_game_leader_selection_dialog_slots.dart';
+part 'new_game_leader_selection_dialog_slot_row.dart';
 part 'new_game_leader_selection_dialog_setup_fields.dart';
 
 const int _kNumSlots = 6;
@@ -119,6 +120,7 @@ class _NewGameLeaderSelectionDialogState extends State<NewGameLeaderSelectionDia
     with
         _NewGameLeaderSelectionDialogStateBase,
         _NewGameLeaderSelectionDialogSlots,
+        _NewGameLeaderSelectionDialogSlotRow,
         _NewGameLeaderSelectionDialogSetupFields {
   @override
   Widget build(BuildContext context) {
